# Improve `trim_transparency` and PNG rasterization