For PNG rasterization, one quick optimization is to do `multiply_alpha` in the same loop assigning each pixel. When the PNG dimension is large the time here is not ignorable.

For `trim_transparency`, with the current solution we always loop over all pixels to find the boundary, so doing it n times will result in a time complexity of `O(n*WH)`.

This PR changes it to find the first non-transparent pixel starting from each edge. For example we loop over the (1) area until reaching a pixel. And then do the same for (2) (3) (4):

![image](https://user-images.githubusercontent.com/3676859/183252404-23832d92-8c73-4200-9d64-d3316b2a2851.png)

Still, no pixel is scanned twice but we are now skipping all the pixels inside the white area. In the worse case (transparent canvas) this algorithm has the same performance as our current one. But in average case this is a lot faster as the complexity becomes `O(n*(WH-s))` where s is the average non-transparent area size.

Since initially we always scan the pixels in the original data storage order, another optimization is to skip all-zero bytes quickly. That's because a transparent pixel is likely to have rgba(0, 0, 0, 0). Here I'm using `u128::from_ne_bytes((&pixels[idx..(idx + 16)]).try_into().unwrap())` to read 16 bytes which is 4 pixels at a time.

I tested the SVG in #371 (2636×4235), and it shows a pretty big improvement:

**Current master branch**

```
FontDB init: 61.98ms
Reading: 0.70ms
Parsing: 38.02ms
Rendering: 18958.88ms
Saving: 222.37ms
./target/release/resvg --perf test.svg test.png  17.42s user 1.81s system 95% cpu 20.207 total
```

**This PR**

```
FontDB init: 61.36ms
Reading: 0.59ms
Parsing: 33.39ms
Rendering: 8535.01ms
Saving: 232.73ms
./target/release/resvg --perf test.svg test.png  7.50s user 1.31s system 99% cpu 8.880 total
```

Which is 2.22x faster in _Rendering_ phase.