@RazrFalcon 2022-08-06
Thanks! Correctness and not performance is the current priority for me, therefore there are a lot of easy optimizations here and there. Clipping and text are the obvious one, but it would take a lot of time to implement. 

As for the patch, I'm a bit skeptical about the use of `u128`. Specifically from the portability standpoint. Will it work on WASM? Does it provide any benefits on 32bit systems? Wouldn't it be easier to use `u64`? Is the performance difference that big?