@shuding 2022-08-06
Thanks for creating this library and taking a look the PR! It makes total sense that correctness is the most important thing. I'm digging into the code and trying to optimize it because I want to make it faster for my use case where both performance and size are critical. Noticed some easy optimizations and thought they're nice to have, as they don't affect the correctness regardless. :)

> As for the patch, I'm a bit skeptical about the use of u128.

Good points. I updated the PR to use `u64` instead, which isn't as fast as `u128` but still ~2x faster.

```
FontDB init: 67.42ms
Reading: 0.64ms
Parsing: 34.19ms
Rendering: 10685.33ms
Saving: 221.08ms
./target/release/resvg --perf test.svg test.png  9.11s user 1.85s system 99% cpu 11.030 total
```