@shuding 2022-08-06
Kinda related, I know that you also have ideas like #146 planned, which is pretty complex and you are definitely busy. So I would like to know that are you open to accept performance optimization PRs like this one? And if so, do you have any suggestions on the direction (like SIMD which is more general)?

For example, I was also thinking about improving the `apply_drop_shadow` process. Right now both `box_blur` and `iir_blur` computes all RGBA channels but for the drop shadow use case, only alpha is needed. A quick test shows a 30% speed improvement for that filter, if we only compute blur for alpha channel.