@RazrFalcon 2022-08-06
Clip path optimizations are endless...

1. Right now, our clip is raster and RGBA, but only Alpha is needed. Which means we should rewrite it to use `tiny_skia::ClipMask` at least. The problem is that `tiny_skia::ClipMask` is very limited at the moment and has to be updated accordingly. 
2. Neither `resvg` nor `tiny-skia` support Rectangle clip optimization, when we do not need a mask at all. This would require some, maybe even significant, changes to `tiny-skia`.
3. We could even port Skia's clipping algorithm, which doesn't use an alpha mask at all. Tons of work and code, but way faster.
4. Remember, that SVG clipping is not just a single path/layer. SVG clipPath can be nested. That's why we're using bitmap to begin with. So while we can use "pure vector clip path" optimization (will still be rasterized at some point) in most cases, there are still cases we're we would need to essentially blend clip paths together. 

As for filters, yeah, they are slow. Not sure if blur can be alpha-only thought.